Agentic Ai/Schedule_gen/pyproject.toml
[project]
name = "schedule-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "pymongo>=4.15.1",
    "python-dotenv>=1.1.1",
    "werkzeug>=3.1.3",
]
